Clean up according to code review
