--- src/Notes.json
@@ -1,8 +1,0 @@
-[
-  {
-    "id": "5oRMd-NPHRLqdEGJ39yCy",
-    "created": "Thu May 14 2020 14:46:00 GMT+0200 (Central European Summer Time)",
-    "set": "default",
-    "text": "This is an example note. Edit me!"
-  }
-]
