+++ src/components/SessionPrompt.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import Prompt from '../components/Prompt'
 import electricLightBulb from '../img/electric_light_bulb.png'
+import prompts from '../prompts.json'
+import { getRandomInt } from '../utils/services'
+import { setToLocalStorage } from '../utils/services'
 import Backdrop from './Backdrop'
 import Icon from './Icon'
 import NavButton from './NavButton'
-import prompts from '../prompts.json'
-import { getRandomInt, setToLocalStorage } from '../utils/utils'
 
 export default function SessionPrompt({
   onClickShuffle,
